Suppress nczarr_test/tst_unknown filter test (and its nczarr_test equivalen)
The test case nc_test4/tst_unknown.sh deletes and then restores
a filter in the plugins directory. The test nczarr_test/run_unknown.sh
also does this.  However if both are running at the same time in a
parallel bit, they apparently can interfere and can cause a race
condition failure.

The solution is to suppress one of them. Since nczarr code
is more unstable, we need to run this test. So suppress the
corresponding test in nc_test4.