CI: Change autotools CI build to out-of-tree build.
This is a reference to an issue with how most distribution packagers
run autotools (source in one directory, compile in another, install
to a third.
There was a PR to catch errors in that kind of build by running
make distcheck; this should do the relevant bits of that PR,
taking into account the preference for separate build and compile
steps.