Modify ncdump to print char-valued variables as utf8.
re: Issue https://github.com/Unidata/netcdf-c/issues/2916

Currently, ncdump prints char-valued variables as a mix
of ascii and octal characters. The octal format is used
for non-printable ascii character values.

This PR changes this to print the char variable values
as raw binary. This means in practice that utf-8 tags
are properly interpreted and printed as utf-8.