Fix some warning in cmake
re: Issue https://github.com/Unidata/netcdf-c/issues/2939

The problem managing inter-dependent options.
I was looking for certain cases of the form, where x and y are options.
````
if( not x and y)
message(warning "not x => not y)
endif
````
However in a couple of cases (including the referenced Issue), I was only testing this:
````
if( not x)
message(warning "not x => not y)
endif
````

Fix is to correct the tests.