Clean up some inconsistencies
There were several HDF5_PLUGIN_PATH that did not have the underscore escaped.  A few other symbols with similar issues.

Fixed some code blocks that appeared broken on my markdown viewer... Not sure if that is the case in general.

There was a HDF5_PLUGIN_DIR referenced near the bottom.  I think it was supposed to be _PATH, but could be wrong since there are multiple _DIR and _PATH througout the code and scripts...