Cleanup the blosc testing in nc_test4 and nczarr_test.
re: Issue https://github.com/Unidata/netcdf-c/issues/2868

There are three parameters (zero, one, and two) that are currently
used in testing, but should not be. Parameters zero and one
are ignored by the code and parameter two is the typesize
parameter that either takes on the value 0 or 4 (== sizeof(int)).
See nc_test4/tst_specific_filters.sh.

Unfortunately, these three parameters differ depending on the blosc library.
The solution is to force them to a specific value -- namely 0,0,0 --
in the test script.