Provide an auxilliary function that allows users to parse the _NCProperties attribute.
re: Discussion https://github.com/Unidata/netcdf-c/discussions/3085

This discussion raised the issue of the best way to distinguish
a netcdfd-c created file and an HDF5 created file. The
recommended way is to use the _NCProperties attribute.  In order
for users to process this attribute, I have added a parser for
the attribute to the netcdf_aux.h file.