Modify provenance check "_IsNetcdf4" to check for existence of "_NCProperties"
re: https://github.com/Unidata/netcdf-c/discussions/3085

This PR adds a check for _NCProperties provenance as part of _IsNetcdf4 test.
This simplifies what the user needs to check to determine if a file is Netcdf-4.
It also reduces the number of false negatives.

The primary change is to libhdf5/nc4hdf4.c. This change changes
NC4_strict_att_exists to NC4_root_att_exists. This is used to
check for root attributes whose existence indicate that a file
is a Netcdf-4 file.