Remove strlcpy from libdispatch
It seems like this function was not used somewhere. Having this symbol
there is causing issues for me a project where I link netcdf and several
other libraries. One of the other libraries (libpq) also provides a `strlcpy`
implementation which results in a duplicated symbol error on linking.

Given that netcdf does not seem to actually use this function I figured
it would be the easiest solution to just remove it here.