Fix the list of libraries expended in the pc file
Modifies the private dependency string created from CMake targets to
ensure that the names of the libraries listed in the generated pc file
are not CMake-defined target names. They must correctly refer to the
libraries on disk, as expected by the "-l" link option.

Note that this issue appears to be recurring and occurs frequently, in
fact every time CMake targets change their naming convention.