Fix bug in ncdump when printing dimension FQN
re: Issue https://github.com/Unidata/netcdf-c/issues/3184

When ncdump attempts to find a dimension's Fully Qualified Name (FQN),
it uses a two step process.
1. Search up the group tree starting at the group in which the dimension the referenced.
2. If not found, then do a breadth first search of the whole group tree looking for the dimension.

Once located, a FQN is constructed with respect to the group in which the dimension was found.

It turns out there was a bug in the ncdump function "searchgroupdim" such that it always returned the top level group as the one containing the dimension.
This, of course is erroneous for nested dimension.

So the "searchgroupdim" function was fixed to properly fail if the dimension was not actually in the root group.