Fix Issue with Numcodecs encoding problems
# re: Issue https://github.com/Unidata/netcdf-c/issues/3200

## Primary Change(s)

The above issue notes that the zarr encoding of some filter parameters is encoding the integer value as a string rather than a proper integer. So this PR fixes the ones identified by the above issue as well as other occurrences of this encoding issue, but not identifed in the above Issue.

The set of files affected by this problem are:
- plugins/NCZhdf5filters.c
- plugins/NCZstdfilters.c
- nc_test4/tst_specific_filter.sh
- nczarr_test/ref_bzip2.cdl
- nczarr_test/ref_filtered.cdl
- nczarr_test/ref_multi.cdl

Note that since the test case nczarr_test/run_specific_filter.sh is actually derived from the file nc_test4/tst_specific_filter.sh, the filter encoding problem must be fixed in the latter file.

## Secondary Issues

There is a backup implementation of memmove in libdispatch/dmissing.c for use in the event that the platform does not support it. I discovered by accident that that implementation was erroneous, so repaired it. The file nclist.c has its own backup memmove implementation and that was repaired also. Also added a declaration for it in include/ncconfigure.h. Also found all uses of memove (esp. libdap4) and fixed those occurrences.

## Misc. Fixes

* Removed some debug output from unit_test/Makefile.am
* Discovered and fixed an error in filter list reclamation in libhdf5/hdf5filter.c
* Fixed an error in d4meta.c where we are setting a list's allocated space instead of its length.
* In some tests, various files copied, unzip'd, untarred, etc. Because these files are shared across subtests, they can cause "file already exists" errors. Solution is to remove those files before copying/unzip'ing into them. Affected filess are:
    - nczarr_test/run_interop.sh
    - nczarr_test/run_notzarr.sh