Modify the way xarray attribute sets are handled.
ref: Issue https://github.com/Unidata/netcdf-c/issues/3214

The xarray named attribute mechanism causes some problems in
NCZarr because it's names are not fully qualified. This means
that whenever a set _ARRAY_ATTRIBUTE dimension names are
encountered, mapping them to/from Netcdf-4 dimensions is a bit
tricky.

Anyway, we think we have been implementing them incorrectly since
we give, e.g. the xarray name "x" the FQN of "/x".  That is, the
dimension is created in the root group.  This is problematic
since the value assigned to x is actually determined by the
shape of the enclosing array but the value assigned to x is not
fixed across the whole dataset.

Since netcdf-4 defines dimensions at the group level, we must
also do so, and that is the primary change in this PR.

Note that this does not solve everything. Consider a Zarr group
with two variables and where each variable has same rank but
different shape values, and where the _ARRAY_ATTRIBUTE list for
each is the same.  Then it is impossible to assign a unique
value to each attribute. In this case, the best we can do (I
think) is fail and suppress the use of _ARRAY_ATTRIBUTE inside
the group.

Also added a test case to check this change and also to verify
that anonymous dim placement is unchanged.

P.S. Note that this PR does not actually fix Issue
https://github.com/Unidata/netcdf-c/issues/3214. The problem in
that Issue is that the zarr '<U1' dtype is being used. This is
the old Python unicode type and we can find no appropriate hack
for this datatype.