Fix H5FD_class_t problems
re: Issue https://github.com/Unidata/netcdf-c/issues/3202

## Primary Change

The notion of versioning H5FD_class_t was introduced publicly in
hdf5-1.13.2. This seems to conflict with some documentation that
says it was 1.10.0. This latter value found its way into
H5FDhttp.c.  As a result, there is a discrepancy in
libhdf5/H5FDhttp.c.  Note that this should address the above
issue in part.

So, the fixes:
1. Regularize this the version testing to be hdf5 version 1.13.2.
2. Use the correct H5FD_class_t fields depending on the hdf5 version.

## Secondary Change
I noted a problem with the definition of strndup when compiling
with C++.  Since strndup is defined for C++, I made the strndup
decl be conditional on not C++.